==> [4] variable creation/step3-Replace NA with 365.ipynb <==
# Replace NA with 365

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score as acc
from mlxtend.feature_selection import SequentialFeatureSelector as sfs
from sklearn.model_selection import train_test_split
from sklearn import metrics

data = pd.read_csv("FinalNewVariables.csv")
#file that after features creation and with NA values, and we will replace NA with 365 now

dataFraudLabel = pd.read_csv("applications data.csv")

dataFraudLabel.head()

dataFraudLNew = dataFraudLabel.drop(columns = ['date','ssn','firstname','lastname','address','zip5','dob','homephone'])

dataFraudLNew.head()

data.isnull().sum()

data.shape

data2 = data.iloc[:,1:]

data2.dtypes

#data2.describe().transpose()

data3 = data2.fillna(365)

new = pd.merge(data3,dataFraudLNew,on = 'record')

new.tail(5)

new.isnull().sum().sum()

new.to_csv("FinalCleanedVariablesNoNa.csv")
